+++ Source/ChilliSource/Rendering/Model/StaticMeshComponent.cpp
@@ -207,7 +207,7 @@ namespace ChilliSource
             if (nullptr != mpModel)
             {
                 s32 indwIndex = mpModel->GetSubMeshIndexByName(instrSubMeshName);
-                if (indwIndex >= 0 && indwIndex < mMaterials.size())
+				if (indwIndex >= 0 && indwIndex < (s32)mMaterials.size())
                 {
                     mMaterials[indwIndex] = inpMaterial;
                     
@@ -240,7 +240,7 @@ namespace ChilliSource
             {
                 s32 indwIndex = mpModel->GetSubMeshIndexByName(instrSubMeshName);
             
-                if (indwIndex >= 0 && indwIndex < mMaterials.size())
+				if (indwIndex >= 0 && indwIndex < (s32)mMaterials.size())
                 {
                     return mMaterials[indwIndex];
                 }
